Edit: Move Map component into separate folder
