+++ src/pages/property/PropertyDetail.js
@@ -15,7 +15,7 @@ import { axiosRes } from "../../api/axiosDefaults";
 import btnStyles from "../../styles/Button.module.css";
 import styles from "../../styles/PropertyDetail.module.css";
 import Avatar from "../../components/avatar/Avatar.js";
-import Map from "../../components/Map";
+import Map from "../../components/map/Map.js";
 
 /**
  * Code adapted from Code Institute's "Moments" walkthrough.
